Convert watchStuckJobs implementation to use SCAN instead of KEYS
Adjusts the implementation of the watchStubJobs LUA script to use
a paginated SCAN of the keyspace, rather than the current KEYS
approach.

This has a significant impact in large or shared redis deployments
where the keyspace can be unbounded in size.  KEYS will block
all execution as it runs, which can take several seconds or longer
whereas the SCAN approach permits other operations to be run in
between each page of results.